+++ packages/web/design-system/src/ui/avatar/avatar/Avatar.vue
@@ -6,6 +6,8 @@ import {
 } from 'reka-ui'
 import { computed } from 'vue'
 
+import type { AvatarColorStyle } from '@/ui/avatar/avatar/avatar.constant'
+import { avatarColorMap } from '@/ui/avatar/avatar/avatar.constant'
 import type { AvatarProps } from '@/ui/avatar/avatar/avatar.props'
 import { AVATAR_DEFAULTS } from '@/ui/avatar/avatar/avatar.props'
 import type { AvatarStyle } from '@/ui/avatar/avatar/avatar.style'
@@ -15,6 +17,17 @@ const props = withDefaults(defineProps<AvatarProps>(), {
   ...AVATAR_DEFAULTS,
 })
 
+const avatarStyle = computed<AvatarStyle>(
+  () => createAvatarStyle({
+    size: props.size,
+    status: props.status ?? undefined,
+  }),
+)
+
+const style = computed<AvatarColorStyle>(
+  () => avatarColorMap[getAvatarIndex(props.name, avatarColorMap.length)]!,
+)
+
 function getInitials(name: string): string {
   const nameParts = name.split(' ')
   const initials = nameParts.map((part) => part.charAt(0).toUpperCase())
@@ -23,41 +36,47 @@ function getInitials(name: string): string {
   return initials.slice(0, count).join('')
 }
 
-const avatarStyle = computed<AvatarStyle>(
-  () => createAvatarStyle({
-    size: props.size,
-    status: props.status ?? undefined,
-  }),
-)
+function getAvatarIndex(name: string, length: number): number {
+  let hash = 0
+
+  for (let i = 0; i < name.length; i++) {
+    hash = (hash << 5) - hash + name.charCodeAt(i)
+    hash |= 0
+  }
+
+  return Math.abs(hash) % length
+}
 </script>
 
 <template>
-  <div :class="avatarStyle.root()">
-    <AvatarRoot
-      :key="props.src ?? 'fallback'"
-      :as-child="true"
+  <AvatarRoot :class="avatarStyle.root()">
+    <AvatarImage
+      v-if="props.src"
+      :src="props.src"
+      :class="avatarStyle.base()"
+      :alt="props.name"
+    />
+
+    <AvatarFallback
+      :class="avatarStyle.fallBack({
+        class: props.isStaticColor
+          ? undefined
+          : `${style.backgroundColor} ${style.textColor}`,
+      })"
     >
-      <AvatarImage
-        v-if="props.src"
-        :src="props.src"
-        :class="avatarStyle.base()"
-        :alt="props.imageAlt ?? 'Avatar'"
-      />
-      <AvatarFallback
-        :class="avatarStyle.fallBack()"
-      >
-        {{ getInitials(props.name) }}
-      </AvatarFallback>
-    </AvatarRoot>
+      {{ getInitials(props.name) }}
+    </AvatarFallback>
+
     <span
-      v-if="props.status != null"
+      v-if="props.status !== null"
       :class="avatarStyle.statusDot()"
     />
+
     <img
       v-else-if="props.logo"
       :src="props.logo"
       :alt="props.logoAlt ?? 'Avatar Logo'"
       :class="avatarStyle.logo()"
     >
-  </div>
+  </AvatarRoot>
 </template>