+++ packages/web/design-system/src/ui/button/icon/stories/IconButtonStatesPlayground.vue
@@ -83,7 +83,10 @@ import IconButton from '@/ui/button/icon/IconButton.vue'
       <div>
         <IconButton
           :icon="Settings01Icon"
-          keyboard-shortcut="⌘,"
+          :keyboard-shortcut="{
+            key: ',',
+            meta: true,
+          }"
           label="Open Settings"
           tooltip-label="Settings"
         />